+++ features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/attachments/preview/AttachmentsPreviewStateProvider.kt
@@ -31,7 +31,6 @@ fun anAttachmentsPreviewState(
 ) = AttachmentsPreviewState(
     attachment = Attachment.Media(
         localMedia = LocalMedia("file://path".toUri(), mediaInfo),
-        compressIfPossible = true
     ),
     sendActionState = sendActionState,
     eventSink = {}